requirements.txt
fastapi
uvicorn
httpx
elasticsearch==8.11.1
python-dotenv
